Refactor navigation store & create auth store
