+++ JoyboyCommunity/src/app/Router.tsx
@@ -5,7 +5,6 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {useMemo} from 'react';
 import {useTheme} from 'styled-components/native';
 
-import useNavigationStore from '../hooks/useNavigationContext';
 import Error from '../modules/error';
 import Feed from '../modules/feed';
 import Login from '../modules/login';
@@ -14,6 +13,7 @@ import CreatePost from '../modules/post';
 import Profile from '../modules/profile';
 import {NoteDetail} from '../screens/NoteDetail';
 import {UserDetail} from '../screens/UserDetail';
+import {useNavigationStore} from '../store/navigation';
 import {HomeStackParams, RootStackParams} from '../types';
 
 const RootStack = createNativeStackNavigator<RootStackParams>();
